JE2-152 1. Back button added to client, product and stock detail pages 2. Delete and Edit buttons moved in components/UI/Buttons folder
