+++ frontend/src/features/stocks/containers/StockDetails.tsx
@@ -1,13 +1,13 @@
-import { Box, Button, CircularProgress, IconButton, Stack, Typography } from '@mui/material'
+import { Box, Button, CircularProgress, Stack, Typography } from '@mui/material'
 import { useStockDetails } from '../hooks/useStockDetails.ts'
 import Modal from '../../../components/UI/Modal/Modal.tsx'
 import StockForm from '../components/StockForm.tsx'
 import Grid from '@mui/material/Grid2'
-import { ArrowBack } from '@mui/icons-material'
 import { DataGrid } from '@mui/x-data-grid'
 import { ruRU } from '@mui/x-data-grid/locales'
-import EditButton from '../../../components/UI/EditButton/EditButton.tsx'
-import DeleteButton from '../../../components/UI/DeleteButton/DeleteButton.tsx'
+import EditButton from '../../../components/UI/Buttons/EditButton.tsx'
+import DeleteButton from '../../../components/UI/Buttons/DeleteButton.tsx'
+import BackButton from '../../../components/UI/Buttons/BackButton.tsx'
 
 const StockDetails = () => {
   const {
@@ -17,7 +17,6 @@ const StockDetails = () => {
     showDeleteModal,
     hideDeleteModal,
     handleDelete,
-    navigateBack,
     editModalOpen,
     setEditModalOpen,
     stockColumns,
@@ -61,17 +60,9 @@ const StockDetails = () => {
         </Grid>
       </Modal>
 
-      <div className="flex items-center gap-3">
-        <IconButton onClick={() => navigateBack()}>
-          <ArrowBack />
-        </IconButton>
-        <Typography variant="h5" fontWeight={700} className="whitespace-normal break-words">
-          Назад
-        </Typography>
-      </div>
-
       <div className="max-w-4xl mx-auto mt-6 bg-white rounded-lg shadow-lg p-8 mb-8">
-        <Box className="text-center mb-8 p-4 bg-gray-100 rounded-lg shadow-md">
+        <BackButton />
+        <Box className="text-center mt-4 mb-8 p-4 bg-gray-100 rounded-lg shadow-md">
           <Typography
             sx={{ fontSize: '20px', fontWeight: 700, color: '#1F2937' }}
             className="whitespace-normal break-words"
